glugun: internal rewrite
